+++ src/dtos/user.dto.ts
@@ -3,5 +3,7 @@ export class CreateUserDto {
     readonly lastName: string;
     readonly nickName: string;
     readonly phoneNumber: string;
+    readonly password: string;
     readonly messageIds: number[];
+    readonly userIds: number[];
 }