+++ src/modules/chat.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
+import { ChatService } from '../services/chat.service';
 
 @Module({
   imports: [],
   controllers: [],
-  providers: [],
+  providers: [ ChatService ],
 })
 export class ChatModule {}