+++ src/modules/message.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
+import { MessageService } from '../services/message.service';
 
 @Module({
   imports: [],
   controllers: [],
-  providers: [],
+  providers: [ MessageService ],
 })
 export class MessageModule {}